+++ buildVars.py
@@ -1,4 +1,4 @@
-from site_scons.site_tools.NVDATool.typings import AddonInfo, BrailleTables, SymbolDictionaries
+from site_scons.site_tools.NVDATool.typings import AddonInfo, BrailleTables, SymbolDictionaries, SpeechDictionaries
 from site_scons.site_tools.NVDATool.utils import _
 
 addon_info = AddonInfo(
@@ -6,14 +6,14 @@ addon_info = AddonInfo(
 	addon_summary=_("Accessible KBBI - Indonesian Dictionary Search"),
 	addon_description=_("""Seamlessly search definitions in the Kamus Besar Bahasa Indonesia (KBBI), the official standard dictionary of the Indonesian language.
 Features include looking up selected text effectively, Word of the Day, and favorite entries, all accessible directly from NVDA."""),
-	addon_version="1.0",
-	addon_changelog=_("""Initial Release."""),
+	addon_version="1.1",
+	addon_changelog=_("""Updated to NVDA 2026.1, and updated addon templates."""),
 	addon_author="Muhammad <muha.aku@gmail.com>",
 	addon_url="https://github.com/muhammadGagah/nvda-accessibleKBBI/",
 	addon_sourceURL="https://github.com/muhammadGagah/nvda-accessibleKBBI/",
 	addon_docFileName="readme.html",
 	addon_minimumNVDAVersion="2024.1",
-	addon_lastTestedNVDAVersion="2025.3.2",
+	addon_lastTestedNVDAVersion="2026.1.1",
 	addon_updateChannel=None,
 	addon_license="GPL-2.0",
 	addon_licenseURL="https://www.gnu.org/licenses/gpl-2.0.html",
@@ -23,7 +23,13 @@ pythonSources: list[str] = ["addon/globalPlugins/accessibleKBBI/*.py", "addon/in
 
 i18nSources: list[str] = pythonSources + ["buildVars.py"]
 
-excludedFiles: list[str] = []
+excludedFiles: list[str] = [
+	"__pycache__/*",
+	"**/__pycache__/*",
+	"*.pyc",
+	"*.pyo",
+	"globalPlugins/accessibleKBBI/accessibleKBBI.json",
+]
 
 baseLanguage: str = "id"
 
@@ -32,3 +38,5 @@ markdownExtensions: list[str] = []
 brailleTables: BrailleTables = {}
 
 symbolDictionaries: SymbolDictionaries = {}
+
+speechDictionaries: SpeechDictionaries = {}